[feat][resotocore] Define a search via command line
