+++ resotocore/resotocore/report/inspector_service.py
@@ -6,7 +6,7 @@ from aiostream import stream
 from attr import evolve, define
 
 from resotocore.analytics import CoreEvent
-from resotocore.cli.cli import CLI
+from resotocore.cli.cli import CLIService
 from resotocore.cli.model import CLIContext
 from resotocore.config import ConfigEntity, ConfigHandler
 from resotocore.db.model import QueryModel
@@ -54,7 +54,7 @@ class CheckContext:
 
 
 class InspectorService(Inspector, Service):
-    def __init__(self, cli: CLI) -> None:
+    def __init__(self, cli: CLIService) -> None:
         self.config_handler: ConfigHandler = cli.dependencies.config_handler
         self.db_access = cli.dependencies.db_access
         self.cli = cli
