+++ resotocore/tests/resotocore/cli/command_test.py
@@ -17,9 +17,9 @@ from pytest import fixture
 
 from resotocore import version
 from resotocore.cli import is_node
-from resotocore.cli.cli import CLI
-from resotocore.cli.command import HttpCommand, JqCommand, WorkerCustomCommand, AggregateCommand
-from resotocore.cli.model import CLIDependencies, CLIContext
+from resotocore.cli.cli import CLIService
+from resotocore.cli.command import HttpCommand, JqCommand, AggregateCommand
+from resotocore.cli.model import CLIDependencies, CLIContext, WorkerCustomCommand, CLI
 from resotocore.cli.tip_of_the_day import generic_tips
 from resotocore.console_renderer import ConsoleRenderer, ConsoleColorSystem
 from resotocore.db.graphdb import ArangoGraphDB
@@ -120,7 +120,7 @@ async def test_descendants(cli: CLI) -> None:
 
 
 @pytest.mark.asyncio
-async def test_search_source(cli: CLI) -> None:
+async def test_search_source(cli: CLIService) -> None:
     result = await cli.execute_cli_command('search is("foo") and some_int==0 --> identifier=~"9_"', stream.list)
     assert len(result[0]) == 10
     await cli.dependencies.template_expander.put_template(
@@ -293,7 +293,7 @@ async def test_format(cli: CLI) -> None:
 
 
 @pytest.mark.asyncio
-async def test_workflows_command(cli: CLI, task_handler: TaskHandlerService, test_workflow: Workflow) -> None:
+async def test_workflows_command(cli: CLIService, task_handler: TaskHandlerService, test_workflow: Workflow) -> None:
     async def execute(cmd: str) -> List[JsonElement]:
         ctx = CLIContext(cli.cli_env)
         return (await cli.execute_cli_command(cmd, stream.list, ctx))[0]  # type: ignore
@@ -331,7 +331,7 @@ async def test_workflows_command(cli: CLI, task_handler: TaskHandlerService, tes
 
 
 @pytest.mark.asyncio
-async def test_jobs_command(cli: CLI, task_handler: TaskHandlerService, job_db: JobDb) -> None:
+async def test_jobs_command(cli: CLIService, task_handler: TaskHandlerService, job_db: JobDb) -> None:
     async def execute(cmd: str) -> List[List[JsonElement]]:
         ctx = CLIContext(cli.cli_env)
         return await cli.execute_cli_command(cmd, stream.list, ctx)
@@ -407,7 +407,7 @@ async def test_jobs_command(cli: CLI, task_handler: TaskHandlerService, job_db: 
 
 @pytest.mark.asyncio
 async def test_tag_command(
-    cli: CLI, performed_by: Dict[str, List[str]], incoming_tasks: List[WorkerTask], caplog: LogCaptureFixture
+    cli: CLIService, performed_by: Dict[str, List[str]], incoming_tasks: List[WorkerTask], caplog: LogCaptureFixture
 ) -> None:
     counter = 0
