+++ resotocore/tests/resotocore/query/template_expander_test.py
@@ -1,11 +1,16 @@
+import re
 from datetime import timedelta
 from typing import Dict, Optional, List
 
 import pytest
 
+from resotocore.cli.model import CLI
+from tests.resotocore.db.entitydb import InMemoryDb
 from resotocore.error import NoSuchTemplateError
-from resotocore.query.model import Template
+from resotocore.model.graph_access import Section
+from resotocore.query.model import Template, PathRoot
 from resotocore.query.template_expander import render_template, TemplateExpanderBase, TemplateExpander
+from resotocore.query.template_expander_service import TemplateExpanderService
 from resotocore.types import Json
 from resotocore.util import utc, from_utc
 
@@ -29,6 +34,10 @@ class InMemoryTemplateExpander(TemplateExpanderBase):
 
     def default_props(self) -> Optional[Json]:
         return self.props
+
+    async def parse_query_from_command_line(self, to_parse: str, on_section: Optional[str], **env: str) -> str:
+        # straight forward implementation: simply strip the "search" command
+        return re.sub("^search\\s+", "", to_parse)
 
 
 @pytest.mark.asyncio
@@ -72,6 +81,24 @@ async def test_add_update_delete_get_list(expander: TemplateExpander) -> None:
     assert (await expander.get_template("albs")).template == "is(aws_alb) and age>{{old}} limit 3"  # type: ignore
     await expander.delete_template("albs")
     assert len(await expander.list_templates()) == 0
+
+
+@pytest.mark.asyncio
+async def test_query_creation(cli: CLI) -> None:
+    expander = TemplateExpanderService(InMemoryDb(Template, lambda k: k.name), cli)
+
+    async def compare_query(query: str, on_section: str, omit_expansion: bool) -> None:
+        direct = await expander.parse_query(query, on_section=on_section, omit_section_expansion=omit_expansion)
+        search = await expander.parse_query(
+            f"search {query}", on_section=on_section, omit_section_expansion=omit_expansion
+        )
+        sd = str(direct)
+        ss = str(search)[0 : len(sd)]  # strip sort and limit generated by the command line parser
+        assert sd == ss
+
+    for section in [*Section.content_ordered, PathRoot]:
+        # await compare_query("is(aws_alb) and age>3d", section, True)
+        await compare_query("is(aws_alb) and age>3d", section, False)
 
 
 def test_render_simple() -> None:
